Refactored pydantic validation and added validation to experiments post
