+++ DashAI/back/api/api_v1/endpoints/datasets.py
@@ -4,49 +4,23 @@ import os
 import shutil
 from typing import Union
 
-import pydantic
 from fastapi import APIRouter, Depends, File, Form, Response, UploadFile, status
-from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import HTTPException
 from sqlalchemy import exc
 from sqlalchemy.orm import Session
 
+from DashAI.back.api.api_v1.schemas.datasets_params import DatasetParams
 from DashAI.back.api.deps import get_db
+from DashAI.back.api.utils import parse_params
 from DashAI.back.core.config import component_registry, settings
 from DashAI.back.database.models import Dataset
 from DashAI.back.dataloaders.classes.dashai_dataset import save_dataset
 from DashAI.back.dataloaders.classes.dataloader import to_dashai_dataset
-from DashAI.back.dataloaders.classes.dataloader_params import DatasetParams
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger(__name__)
 
 router = APIRouter()
-
-
-def parse_params(params):
-    """
-    Parse JSON from string to pydantic model.
-
-    Parameters
-    ----------
-    params : str
-        Stringified JSON with parameters.
-
-    Returns
-    -------
-    BaseModel
-        Pydantic model parsed from Stringified JSON.
-    """
-    try:
-        model = DatasetParams.model_validate_json(params)
-        return model
-    except pydantic.ValidationError as e:
-        log.error(e)
-        raise HTTPException(
-            detail=jsonable_encoder(e.errors()),
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-        ) from e
 
 
 @router.get("/")
@@ -135,7 +109,7 @@ async def upload_dataset(
     JSON
         JSON with the new dataset on the database
     """
-    params = parse_params(params)
+    params = parse_params(DatasetParams, params)
     dataloader = component_registry[params.dataloader]["class"]()
     folder_path = os.path.join(settings.USER_DATASET_PATH, params.dataset_name)
