+++ DashAI/back/api/api_v1/endpoints/experiments.py
@@ -1,12 +1,14 @@
 import logging
 from typing import Union
 
-from fastapi import APIRouter, Depends, Response, status
+from fastapi import APIRouter, Depends, Form, Response, status
 from fastapi.exceptions import HTTPException
 from sqlalchemy import exc
 from sqlalchemy.orm import Session
 
+from DashAI.back.api.api_v1.schemas.experiments_params import ExperimentParams
 from DashAI.back.api.deps import get_db
+from DashAI.back.api.utils import parse_params
 from DashAI.back.database.models import Experiment
 
 logging.basicConfig(level=logging.DEBUG)
@@ -67,9 +69,7 @@ async def get_experiment(experiment_id: int, db: Session = Depends(get_db)):
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def upload_experiment(
-    dataset_id: int,
-    task_name: str,
-    name: str,
+    params: str = Form(),
     db: Session = Depends(get_db),
 ):
     """
@@ -88,8 +88,11 @@ async def upload_experiment(
     JSON
         JSON with the new experiment on the database
     """
+    params = parse_params(ExperimentParams, params)
     try:
-        experiment = Experiment(dataset_id=dataset_id, task_name=task_name, name=name)
+        experiment = Experiment(
+            dataset_id=params.dataset_id, task_name=params.task_name, name=params.name
+        )
         db.add(experiment)
         db.commit()
         db.refresh(experiment)
