+++ DashAI/back/api/api_v1/schemas/experiments_params.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class ExperimentParams(BaseModel):
+    dataset_id: int
+    task_name: str
+    name: str
